Move settings parser declaration to separate header
